requirements.txt
fastapi>=0.110
uvicorn>=0.29
requests>=2.31
PySide6>=6.7
pandas>=2.2
openpyxl>=3.1
PyPDF2>=3.0
PyMuPDF>=1.24
Pillow>=10.0
pdf2image>=1.17
pytesseract>=0.3.10
matplotlib>=3.8
